chore: update deps

+++ config/config.go
@@ -106,10 +106,10 @@ type WebConf struct {
 	BodyLimit int `json:"body_limit"`
 
 	// 黑白名单中间件缓存容量配置, 键生命周期秒数
-	WhitelistLRUCapacity uint64 `json:"whitelist_lru_capacity"`
-	WhitelistLRULifetime uint64 `json:"whitelist_lru_lifetime"`
-	BlacklistLRUCapacity uint64 `json:"blacklist_lru_capacity"`
-	BlacklistLRULifetime uint64 `json:"blacklist_lru_lifetime"`
+	WhitelistLRUCapacity uint32 `json:"whitelist_lru_capacity"`
+	WhitelistLRULifetime uint32 `json:"whitelist_lru_lifetime"`
+	BlacklistLRUCapacity uint32 `json:"blacklist_lru_capacity"`
+	BlacklistLRULifetime uint32 `json:"blacklist_lru_lifetime"`
 
 	CertFile string `json:"-"`
 	KeyFile  string `json:"-"`
